Refactor out platformctl package; rename module controller
