+++ module-operator/controllers/common/platform_utils.go
