+++ module-operator/controllers/modlifecycle/module_lifecycle_controller.go
@@ -9,9 +9,9 @@ import (
 	"time"
 
 	modulesv1alpha1 "github.com/verrazzano/verrazzano-modules/module-operator/apis/platform/v1alpha1"
-	"github.com/verrazzano/verrazzano-modules/module-operator/controllers/platformctrl/common"
-	"github.com/verrazzano/verrazzano-modules/module-operator/controllers/platformctrl/modlifecycle/delegates"
-	"github.com/verrazzano/verrazzano-modules/module-operator/controllers/platformctrl/modlifecycle/reconciler"
+	"github.com/verrazzano/verrazzano-modules/module-operator/controllers/common"
+	"github.com/verrazzano/verrazzano-modules/module-operator/controllers/modlifecycle/delegates"
+	"github.com/verrazzano/verrazzano-modules/module-operator/controllers/modlifecycle/reconciler"
 	controller2 "github.com/verrazzano/verrazzano-modules/module-operator/pkg/controller"
 	"github.com/verrazzano/verrazzano/pkg/log/vzlog"
 	"go.uber.org/zap"
