+++ module-operator/controllers/modlifecycle/reconciler/delegates.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 
 	modulesv1alpha1 "github.com/verrazzano/verrazzano-modules/module-operator/apis/platform/v1alpha1"
-	"github.com/verrazzano/verrazzano-modules/module-operator/controllers/platformctrl/modlifecycle/delegates"
+	"github.com/verrazzano/verrazzano-modules/module-operator/controllers/modlifecycle/delegates"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
