+++ module-operator/controllers/modlifecycle/reconciler/helm_component_reconciler.go
@@ -8,8 +8,8 @@ import (
 	"fmt"
 
 	modulesv1alpha1 "github.com/verrazzano/verrazzano-modules/module-operator/apis/platform/v1alpha1"
-	"github.com/verrazzano/verrazzano-modules/module-operator/controllers/platformctrl/common"
-	"github.com/verrazzano/verrazzano-modules/module-operator/controllers/platformctrl/modlifecycle/delegates"
+	"github.com/verrazzano/verrazzano-modules/module-operator/controllers/common"
+	"github.com/verrazzano/verrazzano-modules/module-operator/controllers/modlifecycle/delegates"
 	"github.com/verrazzano/verrazzano-modules/module-operator/pkg/controller"
 	vzlogInit "github.com/verrazzano/verrazzano/pkg/log"
 	"github.com/verrazzano/verrazzano/pkg/log/vzlog"
